+++ src/evaluation_pro/src/module/density/density_eval.h
@@ -22,17 +22,17 @@ class DensityEval
   std::string evalGlobalNetDensity(DensityNets nets, DensityRegion region, int32_t grid_size);
   std::string evalAllNetDensity(DensityNets nets, DensityRegion region, int32_t grid_size);
 
-  std::string reportMacroDensity(int32_t threshold);
-  std::string reportStdCellDensity(int32_t threshold);
-  std::string reportAllCellDensity(int32_t threshold);
+  std::string reportMacroDensity(float threshold);
+  std::string reportStdCellDensity(float threshold);
+  std::string reportAllCellDensity(float threshold);
 
-  std::string reportMacroPinDensity(int32_t threshold);
-  std::string reportStdCellPinDensity(int32_t threshold);
-  std::string reportAllCellPinDensity(int32_t threshold);
+  std::string reportMacroPinDensity(float threshold);
+  std::string reportStdCellPinDensity(float threshold);
+  std::string reportAllCellPinDensity(float threshold);
 
-  std::string reportLocalNetDensity(int32_t threshold);
-  std::string reportGlobalNetDensity(int32_t threshold);
-  std::string reportAllNetDensity(int32_t threshold);
+  std::string reportLocalNetDensity(float threshold);
+  std::string reportGlobalNetDensity(float threshold);
+  std::string reportAllNetDensity(float threshold);
 
  private:
   std::string evalDensity(DensityCells cells, DensityRegion region, int32_t grid_size, std::string cell_type, std::string output_filename);
